requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data processing
datasets>=2.12.0
Pillow>=9.5.0
numpy>=1.24.0

# Metrics and evaluation
nltk>=3.8.1
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Experiment tracking
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
pytest>=7.3.1

# API and deployment
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
